+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - PIM-3752: Fixed the hard coded entry `Select Job` on import/export creation
 - PIM-3736: Fix wrong count of products in Variant group view
 - PIM-3628: Fixed products not being versioned when modifing a metric, price or media value
+- PIM-3753: Fix completeness filter
 
 ## BC breaks
 - Add a TranslatorInterface argument in MetricType::__construct